Split ReminderForm into multiple files
