+++ components/medication/form/index.ts
@@ -1,0 +1,1 @@
+export { ReminderForm } from "./ReminderForm";
