+++ components/medication/index.ts
@@ -2,5 +2,5 @@ export { IntakeEntry } from "./IntakeEntry";
 export { ReminderCard } from "./ReminderCard";
 export { useReminderDeactivateDialog } from "./ReminderDeactivateDialog";
 export { ReminderDetails } from "./ReminderDetails";
-export { ReminderForm } from "./ReminderForm";
+export { ReminderForm } from "./form";
 export { ReminderHistory } from "./ReminderHistory";
